Move echo functions into their own file
