+++ src/rich_click/__init__.py
@@ -13,8 +13,8 @@ from click import *  # noqa: F401, F403
 from click import command as click_command
 from click import group as click_group
 
-from rich_click import echo, echo_via_pager  # noqa: F401
 from rich_click.rich_command import RichCommand
+from rich_click.rich_echo import echo, echo_via_pager  # noqa: F401
 from rich_click.rich_group import RichGroup
 
 # MyPy does not like star imports. Therefore when we are type checking, we import each individual module
